requirements.txt
# ==========================================================
# NARYAN AI — Python Dependencies
# Install: pip install -r requirements.txt
# ==========================================================

# Web framework
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0

# LLM
groq==0.9.0

# PDF / Document parsing
pdfplumber==0.10.3
PyPDF2==3.0.1
docx2txt==0.8
python-docx==1.1.0
python-pptx==0.6.23

# ML / Vector search
scikit-learn==1.4.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
cachetools==5.3.3

# MySQL
mysql-connector-python==8.3.0
